+++ src/js/language/en_US.js
@@ -323,6 +323,7 @@
             countryNotSupported: 'The country code %s is not supported',
             country: 'Please enter a valid %s',
             countries: {
+                BR: 'Brazilian postal code',
                 CA: 'Canadian postal code',
                 DK: 'Danish postal code',
                 GB: 'United Kingdom postal code',